third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-style-blocked-expected.txt

CONSOLE ERROR: Failed to find a valid digest in the 'integrity' attribute for resource 'http://127.0.0.1:8000/security/subresourceIntegrity/style-1-of-3.css' with computed SHA-256 integrity 'RvLeYLQyPa/ZQk95Rj0XQpfsoBHW9Vrqb3zwo5DScrI='. The resource has been blocked.
CONSOLE ERROR: Failed to find a valid digest in the 'integrity' attribute for resource 'http://127.0.0.1:8000/security/subresourceIntegrity/style-2-of-3.css' with computed SHA-256 integrity 'ReqoU3BDt1yeXAp39aL0TXuAI+RK52lE66N3X3bNETo='. The resource has been blocked.
CONSOLE ERROR: Failed to find a valid digest in the 'integrity' attribute for resource 'http://127.0.0.1:8000/security/subresourceIntegrity/style-3-of-3.css' with computed SHA-256 integrity '5pSfxa36wvmGjFUWRF/HbCLmFSs+n44nodSJe3zY3sU='. The resource has been blocked.
This is a testharness.js-based test.
PASS Subresource integrity blocked style resources
Harness: the test ran to completion.
